FEM-2274 DB and session fixes (#33)
Changed primary key of task. The previous primary key was the content URL. This is not good enough because there may be conflicts. The new primary key is the destination path.
Before adding a task check if a task with the same id exists.

In addition, improved session handling in DefaultDownloader.